requirements.txt
greenlet==3.2.5
pillow==11.3.0
pixelmatch==0.3.0
playwright==1.58.0
pyee==13.0.1
typing_extensions==4.15.0
